Factorise out Microphone module
