+++ plugin.xml
@@ -31,11 +31,6 @@
         <source-file src="src/ios/Diagnostic.m" />
 
 
-        <framework src="AVFoundation.framework" />
-        <config-file target="*-Info.plist" parent="NSMicrophoneUsageDescription">
-            <string>This app requires microphone access to function properly.</string>
-        </config-file>
-
         <framework src="CoreMotion.framework" />
         <config-file target="*-Info.plist" parent="NSMotionUsageDescription">
             <string>This app requires motion detection access to function properly.</string>
@@ -169,6 +164,28 @@
         <header-file src="src/ios/Diagnostic_Notifications.h" />
         <source-file src="src/ios/Diagnostic_Notifications.m" />
         <!--END_MODULE NOTIFICATIONS-->
+
+        <!--BEGIN_MODULE MICROPHONE-->
+        <config-file target="config.xml" parent="/*">
+            <feature name="Diagnostic_Microphone">
+                <param name="ios-package" value="Diagnostic_Microphone" />
+                <param name="onload" value="true" />
+            </feature>
+        </config-file>
+
+        <js-module src="www/ios/diagnostic.microphone.js" name="Diagnostic_Microphone">
+            <merges target="cordova.plugins.diagnostic.microphone" />
+        </js-module>
+
+        <framework src="AVFoundation.framework" />
+
+        <config-file target="*-Info.plist" parent="NSMicrophoneUsageDescription">
+            <string>This app requires microphone access to function properly.</string>
+        </config-file>
+
+        <header-file src="src/ios/Diagnostic_Microphone.h" />
+        <source-file src="src/ios/Diagnostic_Microphone.m" />
+        <!--END_MODULE MICROPHONE-->
     </platform>
 
     <platform name="android">
@@ -279,6 +296,12 @@
 
         <source-file src="src/android/Diagnostic_Notifications.java" target-dir="src/cordova/plugins" />
         <!--END_MODULE NOTIFICATIONS-->
+
+        <!--BEGIN_MODULE MICROPHONE-->
+        <js-module name="Diagnostic_Microphone" src="www/android/diagnostic.microphone.js">
+            <merges target="cordova.plugins.diagnostic.microphone" />
+        </js-module>
+        <!--END_MODULE MICROPHONE-->
     </platform>
 
     <platform name="windows">
