+++ src/ios/Diagnostic.m
@@ -91,78 +91,6 @@ ABAddressBookRef _addressBook;
     }
 }
 
-#pragma mark - Audio
-- (void) isMicrophoneAuthorized: (CDVInvokedUrlCommand*)command
-{
-    [self.commandDelegate runInBackground:^{
-        CDVPluginResult* pluginResult;
-        @try {
-#ifdef __IPHONE_8_0
-            AVAudioSessionRecordPermission recordPermission = [AVAudioSession sharedInstance].recordPermission;
-            
-            if(recordPermission == AVAudioSessionRecordPermissionGranted) {
-                pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsInt:1];
-            }
-            else {
-                pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsInt:0];
-            }
-            [self sendPluginResultBool:recordPermission == AVAudioSessionRecordPermissionGranted :command];
-#else
-            [self sendPluginError:@"Only supported on iOS 8 and higher":command];
-#endif
-        }
-        @catch (NSException *exception) {
-            [self handlePluginException:exception :command];
-        };
-    }];
-}
-
-- (void) getMicrophoneAuthorizationStatus: (CDVInvokedUrlCommand*)command
-{
-    [self.commandDelegate runInBackground:^{
-        @try {
-#ifdef __IPHONE_8_0
-            NSString* status;
-            AVAudioSessionRecordPermission recordPermission = [AVAudioSession sharedInstance].recordPermission;
-            switch(recordPermission){
-                case AVAudioSessionRecordPermissionDenied:
-                    status = AUTHORIZATION_DENIED;
-                    break;
-                case AVAudioSessionRecordPermissionGranted:
-                    status = AUTHORIZATION_GRANTED;
-                    break;
-                case AVAudioSessionRecordPermissionUndetermined:
-                    status = AUTHORIZATION_NOT_DETERMINED;
-                    break;
-            }
-            
-            [self logDebug:[NSString stringWithFormat:@"Microphone authorization status is: %@", status]];
-            [self sendPluginResultString:status:command];
-#else
-            [self sendPluginError:@"Only supported on iOS 8 and higher":command];
-#endif
-        }
-        @catch (NSException *exception) {
-            [self handlePluginException:exception :command];
-        }
-    }];
-}
-
-- (void) requestMicrophoneAuthorization: (CDVInvokedUrlCommand*)command
-{
-    [self.commandDelegate runInBackground:^{
-        @try {
-            [[AVAudioSession sharedInstance] requestRecordPermission:^(BOOL granted) {
-                [self logDebug:[NSString stringWithFormat:@"Has access to microphone: %d", granted]];
-                [self sendPluginResultBool:granted :command];
-            }];
-        }
-        @catch (NSException *exception) {
-            [self handlePluginException:exception :command];
-        }
-    }];
-}
-
 #pragma mark - Address Book (Contacts)
 
 - (void) getAddressBookAuthorizationStatus: (CDVInvokedUrlCommand*)command
